nlw/expert/csharp/Auction/src/RocketseatAuction.API/UseCases/Offers/CreateOffer/CreateOfferUseCase.cs
using RocketseatAuction.API.Comunication.Requests;
using RocketseatAuction.API.Contracts;
using RocketseatAuction.API.Entities;
using RocketseatAuction.API.Services;

namespace RocketseatAuction.API.UseCases.Offers.CreateOffer;

public class CreateOfferUseCase
{
    private readonly ILoggedUser _loggedUser;
    private readonly IOfferRepository _repository;
    public CreateOfferUseCase(ILoggedUser loggedUser, IOfferRepository repository)
    {
        _loggedUser = loggedUser;
        _repository = repository;
    }
    public int Execute(int itemId, RequestCreateOfferJson request)
    {
        var user = _loggedUser.User();

        var offer = new Offer
        {
            Price = request.Price,
            CreatedOn = DateTime.Now,
            ItemId = itemId,
            UserId = user.Id
        };

        _repository.Add(offer);
        return offer.Id;
    }
}
